+++ src/Proxy/PcovProxy.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paraunit\Proxy;
+
+class PcovProxy
+{
+    public function isLoaded(): bool
+    {
+        return extension_loaded('pcov');
+    }
+}