+++ CHANGELOG.md
@@ -8,12 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.0] - TBA
 First stable release. The following changes are in comparison to the previous, unstable release (0.12.3), split into the beta releases that were tagged in the meantime.
 
-## [1.0.0-beta2] - TBA
+## [1.0.0-beta2] - 2020-02-26
 ### Added
  * Add support for PHP 7.4
+ * Add support for Symfony 5
+ * Add support for ext-pcov as a coverage driver [#146](https://github.com/facile-it/paraunit/pull/146)
 
 ### Changed
  * Update PHPStan to 0.12 [#145](https://github.com/facile-it/paraunit/pull/145)
+ * Prefer Pcov or Xdebug over PHPDBG as coverage driver [#146](https://github.com/facile-it/paraunit/pull/146)
 
 ## [1.0.0-beta1] - 2019-04-08
 ### Breaking changes